feat: 메인 페이지 구현 완료
