--- src/constants/my-menu.js
@@ -1,28 +1,0 @@
-const MY_MENU_LIST = [
-  {
-    label: '포인트',
-    imgSrc: 'images/coin.png',
-  },
-  {
-    label: '쿠폰',
-    imgSrc: 'images/coupon.png',
-  },
-  {
-    label: '선물함',
-    imgSrc: 'images/gift.png',
-  },
-  {
-    label: '찜',
-    imgSrc: 'images/heart.png',
-  },
-  {
-    label: '주문내역',
-    imgSrc: 'images/receipt.png',
-  },
-  {
-    label: '리뷰관리',
-    imgSrc: 'images/review.png',
-  },
-];
-
-module.exports = { MY_MENU_LIST };
