+++ src/public/stylesheets/index.css
@@ -1,11 +1,18 @@
-.banner > img {
+.banner-section > img {
   width: 100%;
+  height: 100%;
+  object-fit: cover;
 }
 
 .login-link {
+  margin: 0;
   display: flex;
   align-items: center;
-  padding: 1.5rem 0;
+  padding: 1.5rem 1.25rem;
+}
+
+.login-link:hover {
+  background-color: var(--colors-gray-50);
 }
 
 .login-link span {
@@ -15,7 +22,7 @@
   color: var(--colors-gray-500);
 }
 
-.my-menu {
+.my-menu-list {
   display: flex;
   flex-wrap: wrap;
   list-style: none;
@@ -48,4 +55,81 @@
 .my-menu-item:nth-child(3n + 2) {
   border-left: 2px solid var(--colors-gray-100);
   border-right: 2px solid var(--colors-gray-100);
+}
+
+.ad-section {
+  display: flex;
+  align-items: center;
+  border-top: 1rem solid var(--colors-gray-50);
+  border-bottom: 1rem solid var(--colors-gray-50);
+  padding: 1.25rem 1.5rem;
+
+  font-size: 1.25rem;
+  color: #2f9100;
+}
+
+.ad-section > img {
+  margin-right: 1rem;
+}
+
+.menu-list {
+  margin: 0;
+  padding: 0;
+  list-style: none;
+}
+
+.menu-item {
+  padding: 2rem 1.25rem;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  border-bottom: 2px solid var(--colors-gray-100);
+
+  cursor: pointer;
+}
+
+.menu-item:first-child {
+  border-top: 2px solid var(--colors-gray-100);
+}
+
+.menu-item:hover {
+  background-color: var(--colors-gray-50);
+}
+
+.menu-item-content {
+  flex: 1;
+}
+
+.menu-title-wrap {
+  display: flex;
+  align-items: center;
+  column-gap: 0.5rem;
+  margin-bottom: 0.375rem;
+}
+
+.menu-title {
+  font-size: 1.25rem;
+  line-height: 2rem;
+}
+
+.menu-description {
+  margin: 0;
+  font-size: 1.125rem;
+  color: var(--colors-gray-500);
+}
+
+.menu-title-wrap .badge {
+  display: inline-flex;
+  font-size: 0.75rem;
+  font-weight: 700;
+  color: #ffffff;
+  padding: 0.375rem 0.5rem;
+  border-radius: 9999px;
+  background-color: #ec0d89;
+}
+
+i.chevron-forward {
+  width: 1.5rem;
+  height: 1.5rem;
+  background: url('/images/icons/chevron-forward.svg');
 }
