+++ src/public/stylesheets/mixins/header.css
@@ -3,7 +3,6 @@
   top: 0;
   left: 0;
   right: 0;
-  background-color: #ffffff;
 }
 
 .header-nav {
