+++ src/public/stylesheets/style.css
@@ -25,6 +25,7 @@ body {
     Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
+  background-color: #f5f5f5;
 }
 
 body {
@@ -38,6 +39,9 @@ main {
   max-width: 32rem;
   margin: 0 auto;
   padding-bottom: 1.5rem;
+  background-color: white;
+  box-shadow: rgba(17, 17, 26, 0.05) 0px 4px 16px,
+    rgba(17, 17, 26, 0.05) 0px 8px 32px;
 }
 
 a {
